+++ ft_foreach.c
@@ -6,7 +6,7 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/01/08 15:28:35 by pholster       #+#    #+#                */
-/*   Updated: 2019/01/31 17:40:32 by pholster      ########   odam.nl         */
+/*   Updated: 2019/02/25 13:23:34 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 void	ft_foreach(const int *arr, int length, void (*f)(int))
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (f == NULL || arr == NULL)
 		return ;
-	while (i < length)
+	while (i < (size_t)length)
 	{
 		f(arr[i]);
 		i++;
