+++ ft_strclr.c
@@ -6,7 +6,7 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/01/09 13:58:39 by pholster       #+#    #+#                */
-/*   Updated: 2019/01/31 17:53:08 by pholster      ########   odam.nl         */
+/*   Updated: 2019/02/25 13:11:09 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_strclr(char *str)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (str == NULL)
