+++ ft_striter.c
@@ -6,7 +6,7 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/01/10 11:03:33 by pholster       #+#    #+#                */
-/*   Updated: 2019/02/21 13:26:03 by pholster      ########   odam.nl         */
+/*   Updated: 2019/02/25 13:11:55 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striter(const char *str, void (*f)(char *))
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (f == NULL || str == NULL)
